backend/script/requirements.txt
connexion
Pillow
flask_cors
matplotlib
tk
tensorflow
numpy
pandas